+++ CHANGELOG.md
@@ -7,6 +7,7 @@ UNRELEASED
 - Fix `sos list` command panic if SOS returns bogus entries
 - Fix `lab kube create` node instance upgrade stage (#166)
 - Fix `affinitygroup delete` command confirmation prompt bug (#169)
+- Fix `sos upload` issue with empty files (#173)
 
 1.4.1
 -----